# swagger/api.yaml
openapi: "3.0.0"
info:
  title: BG Library
  version: "1.0"
servers:
  - url: ${API_URL}
    description: Dynamic Server
paths:
...
  /api/v1/library/game:
    post:
      tags:
        - games
      operationId: addGame
      summary: Add a game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '201':
          description: Game added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/library/games:
    get:
      tags:
        - games
      operationId: listGames
      summary: Get list of all games with check out status
      parameters:
        - in: query
          name: checkedOut
          schema:
            type: boolean
          description: Filter games by check out status (optional)
        - in: query
          name: title
          schema:
            type: string
          description: Filter games by title (optional)
      responses:
        '200':
          description: List of games with checkout status successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStatusList'
    post:
      tags:
        - games
      operationId: bulkAddGames
      summary: Add multiple games
      description: |
        Add multiple games by uploading a CSV file. 
        This call is all or nothing. If any game fails to be added, none will be added.
        The CSV file should contain a list of games, one per line, in the following format: 
        "Catan"\n"Ticket to Ride"\n"Azul"
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              format: byte
              description: Base64 encoded CSV of games to add
      responses:
        '201':
          description: Games added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkAddResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/library/game/id/{gameId}:
    get:
      tags:
        - games
      operationId: getGame
      summary: Get a game
      parameters:
        - in: path
          name: gameId
          schema:
            type: string
            format: uuid
          required: true
          description: Game ID
      responses:
        '200':
          description: Game found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - games
      operationId: updateGame
      summary: Update an existing game
      parameters:
        - in: path
          name: gameId
          schema:
            type: string
            format: uuid
          required: true
          description: Game ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '204':
          description: Game successfully updated
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - games
      operationId: deleteGame
      summary: Soft delete a game
      parameters:
        - in: path
          name: gameId
          schema:
            type: string
            format: uuid
          required: true
          description: Game ID
      responses:
        '204':
          description: Game deleted successfully
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/library/game/barcode/{gameBarcode}:
    get:
      tags:
        - games
        - barcode
      operationId: getGameByBarcode
      summary: Get game(s) from a barcode
      description: |
        Get game(s) that match the provided barcode. 
        If multiple games have the same barcode, all will be returned.
        This covers the case where a library uses UPCs in lieu of unique barcodes.
        Unlike the search endpoints, this endpoint will return a 404 if no games are found.
      parameters:
        - in: path
          name: gameBarcode
          schema:
            type: string
          required: true
          description: Game Barcode
      responses:
        '200':
          description: Game(s) found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameList'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    GameStatus:
      type: object
      required:
        - game
      properties:
        game:
          $ref: "#/components/schemas/Game"
        patron:
          $ref: "#/components/schemas/Patron"
        transactionId:
          type: string
          format: uuid
          description: Transaction ID
          example: "f13bb82c-18dd-4a94-a878-44ca6706e2d9"
        checkedOutAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the game was checked out
          example: "2026-01-31T12:00:00Z"
      description: Game with check out info
    Game:
      type: object
      required:
        - gameId
        - title
        - isPlayToWin
      properties:
        gameId:
          type: string
          format: uuid
          description: Game ID
          example: "e1e07b70-c9f3-442f-8225-fe468c3ae705"
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: The name of the board game
          example: "Catan"
        barcode:
          type: string
          minLength: 1
          maxLength: 48
          description: The barcode of the game
          example: "9780307455925"
        isPlayToWin:
          type: boolean
          description: Indicates if the game is a play-to-win game
          example: true
          default: false
      description: A game in the library
    GameList:
      type: object
      required:
        - games
      properties:
        games:
          type: array
          items:
            $ref: '#/components/schemas/Game'
          description: Array of games
      description: List of games
    CreateGameRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: The name of the board game
          example: "Catan"
        barcode:
          type: string
          minLength: 1
          maxLength: 48
          description: The barcode of the game
          example: "9780307455925"
        isPlayToWin:
          type: boolean
          description: Indicates if the game is a play-to-win game
          example: true
          default: false
      description: Request payload for creating a new game
    Patron:
      type: object
      required:
        - patronId
        - name
      properties:
        patronId:
          type: string
          format: uuid
          description: Patron ID
          example: "e1e07b70-c9f3-442f-8225-fe468c3ae705"
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: The name of the patron
          example: "John Smith"
        barcode:
          type: string
          minLength: 1
          maxLength: 48
          description: The barcode of the patron
          example: "9780307455925"
      description: A library patron
...
    GameStatusList:
      type: object
      required:
        - games
      properties:
        games:
          type: array
          items:
            $ref: '#/components/schemas/GameStatus'
          description: Array of the game statuses
      description: List of games with check out info
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - details
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - SERVICE_UNAVAILABLE
                - INTERNAL_ERROR
                - NOT_FOUND
                - CONFLICT
                - MALFORMED_REQUEST
              description: Machine readable error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Validation Error"
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
              description: Array of specific field errors. Empty if no specific errors
      description: Response when errors occur
    ErrorDetail:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          description: The input field that caused the error
          example: "title"
        message:
          type: string
          description: Human-readable error message for this field
          example: "Title cannot exceed 100 characters"
      description: Details about specific errors
    BulkAddResponse:
      type: object
      required:
        - imported
      properties:
        imported:
          type: integer
          description: Number of records imported successfully
          example: 10
      description: Response for bulk add operations
...